Verify the Returncode of Checks
This patch introduces the verification of returncodes of checks.
Despite the fact, that the called tool itself, could have a returncode
unequal 0 and the check still works as expected, there are some
returncodes like 126 and 127, which indicate that the binary was not
properly executed. So this patch checks for those returncodes.